Bump version to flyway-11.3.2
Please see the GH release for the release notes

Fixed an issue where the Flyway 'configuration' API returned Configuration objects with shared references instead of deep copies

Native Connectors now supports transactions for MongoDB with `.json` migrations

+++ flyway-docker/scripts/build_images.py
@@ -45,22 +45,26 @@ def get_build_command(edition, version, tag_suffix, folder, mongo=False):
 if __name__ == "__main__":
     edition = sys.argv[1]
     version = sys.argv[2]
-    
+    linux_only = sys.argv[3] == "True"
     commands = []
-    if edition == "flyway":  # We only do multi-arch builds for OSS due to compatibility issues with Redgate Compare
-        subprocess.run("docker buildx rm multi_arch_builder", shell=True)
-        commands.append("docker run --rm --privileged multiarch/qemu-user-static --reset -p yes")
-        commands.append("docker buildx create --name multi_arch_builder --driver docker-container --driver-opt network=bridge --use")
-        commands.append(get_buildx_command(edition, version, "", "."))
-        commands.append(get_buildx_command(edition, version, "", ".", True))
-        commands.append(get_buildx_command(edition, version, "-alpine", "alpine"))
-        commands.append(get_build_command(edition, version, "-azure", "azure"))
-        commands.append(get_buildx_command(edition, version, "-alpine", "alpine", True))
-        commands.append(get_build_command(edition, version, "-azure", "azure", True))
-    else:
+
+    if linux_only:
         commands.append(get_build_command(edition, version, "", "."))
-        commands.append(get_build_command(edition, version, "-alpine", "alpine"))
-        commands.append(get_build_command(edition, version, "-azure", "azure"))
+    else:
+        if edition == "flyway":  # We only do multi-arch builds for OSS due to compatibility issues with Redgate Compare
+            subprocess.run("docker buildx rm multi_arch_builder", shell=True)
+            commands.append("docker run --rm --privileged multiarch/qemu-user-static --reset -p yes")
+            commands.append("docker buildx create --name multi_arch_builder --driver docker-container --driver-opt network=bridge --use")
+            commands.append(get_buildx_command(edition, version, "", "."))
+            commands.append(get_buildx_command(edition, version, "", ".", True))
+            commands.append(get_buildx_command(edition, version, "-alpine", "alpine"))
+            commands.append(get_build_command(edition, version, "-azure", "azure"))
+            commands.append(get_buildx_command(edition, version, "-alpine", "alpine", True))
+            commands.append(get_build_command(edition, version, "-azure", "azure", True))
+        else:
+            commands.append(get_build_command(edition, version, "", "."))
+            commands.append(get_build_command(edition, version, "-alpine", "alpine"))
+            commands.append(get_build_command(edition, version, "-azure", "azure"))
     
     for command in commands:
         print(f'Running docker build command: {command}')
